requirements.txt
aiofiles==0.4.0
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.4
blinker==1.4
blis==0.2.4
Bottleneck==1.2.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
fastai==1.0.50.post1
fastprogress==0.1.20
flake8==3.7.7
h11==0.8.1
h2==3.1.0
hpack==3.0.0
Hypercorn==0.5.3
hyperframe==5.2.0
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==6.0.0
multidict==4.5.2
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.2
nvidia-ml-py3==7.352.0
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.1.0
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
pynvx==1.0.0
pyparsing==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
pytoml==0.1.20
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
Quart==0.8.1
requests==2.21.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sortedcontainers==2.1.0
soupsieve==1.8
spacy==2.1.0
srsly==0.0.5
terminado==0.8.2
testpath==0.4.2
thinc==7.0.4
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.24.2
wasabi==0.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wsproto==0.13.0
